Split examples
