+++ examples/pliers.py
@@ -2,24 +2,6 @@ import numpy as np
 from mapper_src.mapper import Mapper
 
 
-def noisy_circle(points=100, noise=0.1):
-    """
-    Generates points of a noisy circle (with radius 1).
-    """
-
-    # Generate circle coordinates.
-    d = np.linspace(0, 2 * np.pi, points, endpoint=False)
-    x = np.cos(d)
-    y = np.sin(d)
-
-    # Add Gaussian noise.
-    x = np.random.normal(x, noise)
-    y = np.random.normal(y, noise)
-
-    return np.vstack((x, y)).T
-
-vertices = noisy_circle()
-
 with open('pliers.txt') as f:
     data = f.readlines()
